feat: setBfree organ integration, FX chain overhaul, CC remap
Major changes:
- Integrate setBfree as git submodule for Hammond B3 organ (PG4)
- Add vintage character CC controls: key click (CC21), generator
  leakage (CC22), tone aging (CC23) with serial contact switching
- Remap FxChain CCs to 100+ range to avoid setBfree conflicts
  (drawbars CC70-78, percussion CC80-83, etc.)
- Add IR convolver with Twin Reverb cab IRs (10 mic positions)
- Add LV2 plugin chain: Channel9 Neve + EQ + ChorusEnsemble +
  TapeDelay2 + kPlateB with pre/post IR routing
- Add internal phaser (Small Stone style) with CC3 mod mode switch
- Implement pitch bend (±2 semitones)
- Fix LA Custom XOR decryption and sample parsing
- Convert in-memory PCM to int16 (50% RAM reduction)
- iOS remote: add Organ group, IR group, update CC mappings

+++ src/sample_db.hpp
@@ -3,6 +3,7 @@
 #include <vector>
 #include <map>
 #include <cstddef>
+#include <cstdint>
 
 struct SampleData {
     int         midi_note;
@@ -12,8 +13,9 @@ struct SampleData {
     int         round_robin;
     std::string file_path;
 
-    // RAM に展開した PCM サンプル (float 正規化済み, -1.0 〜 +1.0, モノにダウンミックス済み)
-    std::vector<float> pcm;
+    // RAM に展開した PCM サンプル (int16 正規化済み, モノにダウンミックス済み)
+    // メモリ節約のため int16 で保持し、再生時に float に変換する
+    std::vector<int16_t> pcm;
     int sample_rate  = 44100;
     int num_channels = 1;
 };
